+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Focus highlighting for "Create Group" and "Search Group"
 - Desktop/Web: show scrollbar in user search when creating a new chat
+- Fix crash due to intrinsic size in timeline
 
 ### Security